moved universal time block interfaces too the entities were they belong.
