+++ src/entities/TimeBlock.java
@@ -2,8 +2,6 @@
 
 package entities;
 
-import dataBundles.UniversalTimeBlockWithDay;
-
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
