+++ src/entities/UniversalTimeBlockWithDay.java
@@ -1,4 +1,4 @@
-package dataBundles;
+package entities;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
